Processor object patch.
I've have a patch [1] that uses a much lighter Processor object, while
broadly keeping in line with the previous integration points. It is a
little more invasive therefore I wanted to give folks a chance to look
at it - particularly the integration points before I hit commit. There
is scope to make this cleaner and lighter in Tomcat 8 but that requires
more refactoring of existing code than I really want to do in 7.0.x.